move functions to dataset_and_group_handling.py module
